Update Podcast model and parse YouTube channel ID
The fields in the Podcast model have been updated in the Prisma schema, with the addition of the 'youtubeChannelId' field. If a podcast is being created with a feed URL from YouTube, the 'youtubeChannelId' is parsed and added to the request during podcast creation. Also, additional dependencies have been added to package.json.